Standardize module styling
