+++ src/AppBar/index.jsx
@@ -1,3 +1,3 @@
 module.exports =  {
-    AppBar: require('./Navigation')
+  AppBar: require('./Navigation')
 }
