+++ src/Footer/index.es6
@@ -1,3 +1,1 @@
-const {Footer} = require('./Footer');
-
-module.exports = {Footer};
+module.exports = require('./Footer');
