+++ src/Modal/index.es6
@@ -1,3 +1,1 @@
-const {Modal} = require('./Modal');
-
-module.exports = {Modal};
+module.exports = require('./Modal');
