+++ src/OpenSessionOverview/index.es6
@@ -1,3 +1,1 @@
-const {OpenSessionOverview} = require('./OpenSessionOverview');
-
-module.exports = {OpenSessionOverview}
+module.exports =  require('./OpenSessionOverview');
