+++ src/ProxyPathUtils.es6
@@ -1,24 +1,27 @@
 
-function pathWithSlug(path) { 
+function pathWithSlug(path) {
     var mpath = path;
     if (window.__env && window.__env.config) {
         mpath = (window.__env.config.app.slug || '') + path;
     }
-    return mpath;   
+    return mpath;
 }
 
-function pathWithoutSlug(path) { 
+function pathWithoutSlug(path) {
     var mpath = path;
-    var slug = '' 
+    var slug = ''
     if (window.__env && window.__env.config) {
         slug = window.__env.config.app.slug || '';
     }
 
     if (path.indexOf(slug) == 0) {
-        mpath = path.slice(slug.length);   
+        mpath = path.slice(slug.length);
     }
 
-    return mpath;   
+    return mpath;
 }
 
-module.exports = {pathWithSlug, pathWithoutSlug};
+module.exports = {
+  pathWithoutSlug,
+  pathWithSlug,
+};
