+++ src/SearchBar/index.es6
@@ -1,4 +1,4 @@
-const {SearchActions} = require('./SearchActions');
-const {SearchBar} = require('./SearchBar');
+const SearchActions = require('./SearchActions');
+const SearchBar = require('./SearchBar');
 
-module.exports = {SearchActions, SearchBar}
+module.exports = { SearchActions, SearchBar }
