+++ src/index.es6
@@ -1,25 +1,26 @@
-const {AppBar} = require('./AppBar');
-const {AvailabilityGrid} = require('./AvailabilityGrid');
-const {AvatarUploader} = require('./AvatarUploader');
-const {DatePicker} = require('./DatePicker');
-const {Dropdown} = require('./Dropdown');
-const {Footer} = require('./Footer');
-const {FourOhFour} = require('./FourOhFour');
-const {Gravatar} = require('./Gravatar');
-const {Icon} = require('./Icon');
-const {Loader} = require('./Loader');
-const {Modal} = require('./Modal');
-const {NotificationView} = require('./NotificationView');
-const {OneClickCopy} = require('./OneClickCopy');
-const {OpenSessionOverview} = require('./OpenSessionOverview');
-const {pathWithSlug, pathWithoutSlug} = require('./ProxyPathUtils');
-const {SearchBar} = require('./SearchBar');
-const {SidebarLayout, SidebarMenu, SidebarMenuItem} = require('./Sidebar');
-const {SocialShare} = require('./SocialShare');
-const {Tag} = require('./Tag');
-const {TopicPicker} = require('./TopicPicker');
-const {TrackedLink} = require('./TrackedLink');
-const {models} = require('./models');
+const AppBar = require('./AppBar');
+const AvailabilityGrid = require('./AvailabilityGrid');
+const AvatarUploader = require('./AvatarUploader');
+const DatePicker = require('./DatePicker');
+const Dropdown = require('./Dropdown');
+const Footer = require('./Footer');
+const FourOhFour = require('./FourOhFour');
+const Gravatar = require('./Gravatar');
+const Icon = require('./Icon');
+const Loader = require('./Loader');
+const Modal = require('./Modal');
+const NotificationView = require('./NotificationView');
+const OneClickCopy = require('./OneClickCopy');
+const OpenSessionOverview = require('./OpenSessionOverview');
+const SearchBar = require('./SearchBar');
+const SocialShare = require('./SocialShare');
+const Tag = require('./Tag');
+const TopicPicker = require('./TopicPicker');
+const TrackedLink = require('./TrackedLink');
+
+const models = require('./models');
+const { pathWithSlug, pathWithoutSlug } = require('./ProxyPathUtils');
+const { SidebarLayout, SidebarMenu, SidebarMenuItem } = require('./Sidebar');
 
 module.exports = {
   AppBar,
